+++ owncloudApp/src/main/java/com/owncloud/android/presentation/spaces/SpacesListFragment.kt
@@ -73,6 +73,7 @@ import com.owncloud.android.utils.DisplayUtils
 import com.owncloud.android.presentation.common.UIResult
 import com.owncloud.android.presentation.spaces.createspace.CreateSpaceDialogFragment
 import com.owncloud.android.presentation.transfers.TransfersViewModel
+import com.owncloud.android.presentation.spaces.setspaceicon.SetSpaceIconDialogFragment
 import kotlinx.coroutines.flow.SharedFlow
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import org.koin.core.parameter.parametersOf
@@ -82,7 +83,8 @@ class SpacesListFragment :
     SpacesListAdapter.SpacesListAdapterListener,
     Fragment(),
     SearchView.OnQueryTextListener,
-    CreateSpaceDialogFragment.CreateSpaceListener
+    CreateSpaceDialogFragment.CreateSpaceListener,
+    SetSpaceIconDialogFragment.SetIconListener
 {
     private var _binding: SpacesListFragmentBinding? = null
     private val binding get() = _binding!!
@@ -90,8 +92,7 @@ class SpacesListFragment :
     private var isMultiPersonal = false
     private var userPermissions = mutableSetOf<UserPermissions>()
     private var editQuotaPermission = false
-    private var lastUpdatedRemotePath: String? = null
-    private var selectedImageName: String? = null
+    private var selectedImagePath: String? = null
     private lateinit var currentSpace: OCSpace
 
     private val spacesListViewModel: SpacesListViewModel by viewModel {
@@ -114,7 +115,7 @@ class SpacesListFragment :
             val selectedImageUri = result.data?.data ?: return@registerForActivityResult
             val accountName = requireArguments().getString(BUNDLE_ACCOUNT_NAME) ?: return@registerForActivityResult
             val documentFile = DocumentFile.fromSingleUri(requireContext(), selectedImageUri) ?: return@registerForActivityResult
-            selectedImageName = documentFile.name
+            selectedImagePath = SPACE_CONFIG_DIR + documentFile.name
 
             transfersViewModel.uploadFilesFromContentUri(
                 accountName = accountName,
@@ -247,14 +248,15 @@ class SpacesListFragment :
         }
 
         collectLatestLifecycleFlow(transfersViewModel.transfersWithSpaceStateFlow) { transfersWithSpace ->
-            val remotePath = SPACE_CONFIG_DIR + selectedImageName
-            val matchedTransfer = transfersWithSpace.map { it.first }.find { it.remotePath == remotePath }
+            val matchedTransfer = transfersWithSpace.map { it.first }.find { it.remotePath == selectedImagePath }
 
-            if (matchedTransfer != null && lastUpdatedRemotePath != matchedTransfer.remotePath) {
+            matchedTransfer?.let {
                 when(matchedTransfer.status) {
                     TransferStatus.TRANSFER_SUCCEEDED -> {
                         spacesListViewModel.editSpaceImage(currentSpace.id, matchedTransfer.remotePath)
-                        lastUpdatedRemotePath = matchedTransfer.remotePath
+                        matchedTransfer.id?.let {
+                            transfersViewModel.clearSuccessfulTransferById(it)
+                        }
                     }
                     TransferStatus.TRANSFER_FAILED -> {
                         showMessageInSnackbar(getString(R.string.edit_space_image_failed))
@@ -326,6 +328,17 @@ class SpacesListFragment :
 
     override fun editSpace(spaceId: String, spaceName: String, spaceSubtitle: String, spaceQuota: Long?) {
         spacesListViewModel.editSpace(spaceId, spaceName, spaceSubtitle, spaceQuota)
+    }
+
+    override fun setSpaceIcon(fileName: String, localPath: String) {
+        selectedImagePath = SPACE_CONFIG_DIR + fileName
+        transfersViewModel.uploadFilesFromSystem(
+            accountName = requireArguments().getString(BUNDLE_ACCOUNT_NAME) ?: "",
+            listOfLocalPaths = listOf(localPath),
+            uploadFolderPath = SPACE_CONFIG_DIR,
+            spaceId = currentSpace.id,
+            forceOverwrite = true
+        )
     }
 
     fun setSearchListener(searchView: SearchView) {
@@ -444,6 +457,10 @@ class SpacesListFragment :
                         }
                         editSpaceImageLauncher.launch(action)
                     }
+                    SpaceMenuOption.SET_ICON -> {
+                        val setIconDialog = SetSpaceIconDialogFragment.newInstance(listener = this@SpacesListFragment)
+                        setIconDialog.show(requireActivity().supportFragmentManager, DIALOG_SET_ICON)
+                    }
                 }
             }
         }
@@ -467,6 +484,7 @@ class SpacesListFragment :
         const val SPACE_CONFIG_DIR = "/.space/"
 
         private const val DIALOG_CREATE_SPACE = "DIALOG_CREATE_SPACE"
+        private const val DIALOG_SET_ICON = "DIALOG_SET_ICON"
 
         fun newInstance(
             showPersonalSpace: Boolean,